moved logic to data source
